Code refactoring logging feature. All log config is on config.logger.js. Adding bunyan rolling file stream to rotate log files. Updating .gitignore and dependencies.
