+++ .gitignore
@@ -3,4 +3,4 @@ build/
 .env
 node_modules/
 npm-debug.log
-*.log
+logs/
