+++ package.json
@@ -25,6 +25,7 @@
   "dependencies": {
     "bluebird": "~3.4.3",
     "bunyan": "^1.8.1",
+    "bunyan-rotating-file-stream": "^1.6.1",
     "dotenv": "~2.0.0",
     "lodash": "~4.15.0",
     "mongodb": "~2.2.9",
